filterWorks: Add deferred promise, Add route /work
The filterWorks service now defers the promise and has become
async to allow logic to be written even after the result has been
found. For example, custom errors, etc.

The route /work has been added, it is meant to handle the display
of single works. The corresponding view and controller are also
available. if >1 or 0 results are got, a 404 is thrown.